projects/backend/services/telemetry-ingest-service/pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "telemetry-ingest-service"
version = "0.1.0"
description = "Telemetry ingest service for Experiment Tracking Platform"
authors = ["Product Dev Course"]
readme = "README.md"
packages = [
    { include = "telemetry_ingest_service", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.13.3"
aiohttp-cors = "^0.7.0"
asyncpg = "^0.31.0"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
orjson = "^3.10.3"
structlog = "^24.2.0"
backend-common = { path = "../../common", develop = true }
prometheus-client = "^0.24.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.21.1"
pytest-aiohttp = "^1.0.5"
ruff = "^0.5.0"
mypy = "^1.10.0"
yandex-taxi-testsuite = {extras = ["postgresql-binary"], version = "^0.3.9"}
openapi-generator-cli = "^7.17.0"
watchfiles = "^0.21.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-p testsuite.pytest_plugin"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "ignore:'asyncio\\.iscoroutinefunction' is deprecated.*:DeprecationWarning:pytest_asyncio\\.plugin",
    "ignore:'asyncio\\.get_event_loop_policy' is deprecated.*:DeprecationWarning:pytest_asyncio\\.plugin",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
mypy_path = "../../common/types"

[[tool.mypy.overrides]]
module = "aiohttp_cors.*"
ignore_missing_imports = true

